+++ sample-application/src/main/java/com/alternativepayments/controller/PlanController.java
@@ -1,7 +1,5 @@
 package com.alternativepayments.controller;
 
-import java.math.BigDecimal;
-
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -30,7 +28,7 @@ public class PlanController {
      */
     @RequestMapping("/create-plan")
     public String createPlan(Model model) {
-        Plan plan = new Plan.Builder("Test", BigDecimal.valueOf(1000), "EUR", 5, Plan.Period.DAY)
+        Plan plan = new Plan.Builder("Test", 1000, "EUR", 5, Plan.Period.DAY)
                 .description("Test plan").build();
 
         Plan createdPlan = alternativePaymentClient.create(plan, Plan.API_ENDPOINT, Plan.class);
